group users by age and print results. Fixed
